Revise Dockerfile to copy specific files
The Dockerfile is modified to specifically copy source files, setup files, and the README instead of copying all files. This change assures only necessary files are included and in turn, better structures the docker container. It further allows for more controlled and efficient management of the contents within the docker container.